split resolvers into separate files
